\name{guyton.descs}
\docType{data}
\alias{guyton.descs}
\title{Descriptions of the parameters and variables in the Guyton model}
\description{
  This list contains a description of every parameter and variable in
  the Guyton model. A convenience function, GetDescription(), is
  provided to retrieve descriptions by parameter/variable name.
}
\usage{data(guyton.descs)}
\format{A list containing descriptions of the 410 parameters and
  variables, tagged by name.}
\keyword{datasets}
