\name{qalh.mdflw}
\docType{data}
\alias{qalh.mdflw}
\title{Time-series for Qalh (Moore94) and MDFLW (Guyton92).}
\description{
  This data set contains time-series for the absolute flow at the macula
  densa as predicted by the Moore94 model (Qalh), and for the normalised
  macula densa flow as predicted by the Guyton92 model (MDFLW). By
  scaling Qalh appropriately, it is possible to compare the difference
  in normalised flows between the two models.
}
\usage{data(qalh.mdflw)}
\format{A list of 18,828 pairs of points.}
\keyword{datasets}
\references{
  L. C. Moore et al. Bull Math Biol 56(3):391--410, 1994.
  DOI: \href{http://dx.doi.org/10.1007/BF02460464}{10.1007/BF02460464}.

  A. C. Guyton. Hypertension 19(Suppl 1):I2--8, 1992.
}
