\name{smallppn.ht}
\docType{data}
\alias{smallppn.ht}
\title{A small population of hypertensive individuals}
\description{
  The hypertensive individuals in the randomly-chosen sample population
  "smallppn", where hypertension was defined as having a mean arterial
  pressure greater than (or equal to) 106.6 mmHg.
}
\usage{data(smallppn.ht)}
\format{A data frame containing 653 individuals (rows). Each individual
  is a collection of parameters (columns prefixed by "p_") and
  steady-state variables (columns prefixed by "v_"). Descriptions of the
  parameters and variables are stored in "guyton.descs". Each pair of
  rows were obtained from a delta-perturbation simulation; the even rows
  are the result of perturbing a single parameter from the preceeding
  (odd) rows.}
\keyword{datasets}
