\name{smallppn.nt}
\docType{data}
\alias{smallppn.nt}
\title{A small population of normotensive individuals}
\description{
  The normotensive individuals in the randomly-chosen sample population
  "smallppn", where normotension was defined as having a mean arterial
  pressure less than 93.3 mmHg.
}
\usage{data(smallppn.nt)}
\format{A data frame containing 902 individuals (rows). Each individual
  is a collection of parameters (columns prefixed by "p_") and
  steady-state variables (columns prefixed by "v_"). Descriptions of the
  parameters and variables are stored in "guyton.descs". Each pair of
  rows were obtained from a delta-perturbation simulation; the even rows
  are the result of perturbing a single parameter from the preceeding
  (odd) rows.}
\keyword{datasets}
