\name{smallppn.pht}
\docType{data}
\alias{smallppn.pht}
\title{A small population of pre-hypertensive individuals}
\description{
  The pre-hypertensive individuals in the randomly-chosen sample
  population "smallppn", where pre-hypertension was defined as having a
  mean arterial pressure between 93.3 and 106.6 mmHg.
}
\usage{data(smallppn.pht)}
\format{A data frame containing 725 individuals (rows). Each individual
  is a collection of parameters (columns prefixed by "p_") and
  steady-state variables (columns prefixed by "v_"). Descriptions of the
  parameters and variables are stored in "guyton.descs". Each pair of
  rows were obtained from a delta-perturbation simulation; the even rows
  are the result of perturbing a single parameter from the preceeding
  (odd) rows.}
\keyword{datasets}
